[zero] refactor low-level zero folder structure
