+++ colossalai/zero/legacy/shard_utils/base_shard_strategy.py
@@ -2,6 +2,7 @@ from abc import ABC, abstractmethod
 from typing import List, Optional
 
 import torch.distributed as dist
+
 from colossalai.zero.sharded_param.sharded_tensor import ShardedTensor
