+++ colossalai/zero/legacy/sharded_optim/__init__.py
@@ -1,0 +1,3 @@
+from .sharded_optim_v2 import ShardedOptimizerV2
+
+__all__ = ['ShardedOptimizerV2']
