+++ colossalai/zero/legacy/sharded_param/__init__.py
@@ -1,4 +1,4 @@
-from colossalai.zero.sharded_param.sharded_tensor import ShardedTensor
 from colossalai.zero.sharded_param.sharded_param import ShardedParamV2
+from colossalai.zero.sharded_param.sharded_tensor import ShardedTensor
 
 __all__ = ['ShardedTensor', 'ShardedParamV2']
