+++ colossalai/zero/legacy/sharded_param/sharded_tensor.py
@@ -1,4 +1,5 @@
 import torch
+
 from colossalai.gemini.stateful_tensor import StatefulTensor, TensorState
