--- colossalai/zero/sharded_optim/__init__.py
@@ -1,4 +1,0 @@
-from .low_level_optim import LowLevelZeroOptimizer
-from .sharded_optim_v2 import ShardedOptimizerV2
-
-__all__ = ['ShardedOptimizerV2', 'LowLevelZeroOptimizer']
