+++ tests/test_moe/test_moe_zero_init.py
@@ -1,21 +1,19 @@
 from functools import partial
 
-import colossalai
 import pytest
 import torch
 import torch.multiprocessing as mp
 import torch.nn as nn
-from colossalai.nn import CheckpointModule
-from colossalai.logging import get_dist_logger
-from colossalai.testing import parameterize
-from colossalai.utils import free_port
-from colossalai.context import MOE_CONTEXT
-from colossalai.nn.layer import MoeModule
-from colossalai.zero.init_ctx import ZeroInitContext
-from colossalai.zero.shard_utils import (BucketTensorShardStrategy, TensorShardStrategy)
 
-from colossalai.testing import rerun_if_address_is_in_use
-from colossalai.utils import get_current_device
+import colossalai
+from colossalai.context import MOE_CONTEXT
+from colossalai.logging import get_dist_logger
+from colossalai.nn import CheckpointModule
+from colossalai.nn.layer import MoeModule
+from colossalai.testing import parameterize, rerun_if_address_is_in_use
+from colossalai.utils import free_port, get_current_device
+from colossalai.zero.legacy.init_ctx import ZeroInitContext
+from colossalai.zero.legacy.shard_utils import BucketTensorShardStrategy, TensorShardStrategy
 from tests.test_zero.common import CONFIG
